+++ src/Settings.php
@@ -16,9 +16,16 @@ class Settings {
 	/**
 	 * Hold plugin options.
 	 *
-	 * @var $options
+	 * @var array $options
 	 */
 	protected static $options;
+
+	/**
+	 * Hold non-constant options.
+	 *
+	 * @var array $not_constants
+	 */
+	private static $not_constants = [ 'bypass_shutdown_handler' ];
 
 	/**
 	 * Constructor.
@@ -102,7 +109,9 @@ class Settings {
 	 */
 	private function update_constants( $old, $new ) {
 		$remove = array_diff_assoc( $old, $new );
-		$add    = array_diff_assoc( $new, $old );
+		$remove = $this->filter_constants( $remove );
+		$add    = array_diff_assoc( $new, $old, [ 'shutdown_handler' ] );
+		$add    = $this->filter_constants( $add );
 
 		if ( ! empty( $add ) ) {
 			$this->add_constants( $add );
@@ -110,6 +119,29 @@ class Settings {
 		if ( ! empty( $remove ) ) {
 			$this->remove_constants( $remove );
 		}
+	}
+
+	/**
+	 * Filter and return constants.
+	 *
+	 * Unset saved options that are not constants.
+	 *
+	 * @param array $constants Options.
+	 *
+	 * @return array $constants
+	 */
+	private function filter_constants( $constants ) {
+		if ( empty( $constants ) ) {
+			return $constants;
+		}
+		foreach ( self::$not_constants as $not_constant ) {
+			unset( $constants[ $not_constant ] );
+			$constants = array_flip( $constants );
+			unset( $constants[ $not_constant ] );
+			$constants = array_flip( $constants );
+		}
+
+		return $constants;
 	}
 
 	/**
@@ -234,6 +266,18 @@ class Settings {
 				'title' => esc_html__( 'Set WP_DEBUG_DISPLAY to false, default is true.', 'wp-debugging' ),
 			]
 		);
+
+		add_settings_field(
+			'shutdown_hander',
+			null,
+			[ $this, 'checkbox_setting' ],
+			'wp_debugging',
+			'wp_debugging',
+			[
+				'id'    => 'bypass_shutdown_handler',
+				'title' => esc_html__( 'Bypass WordPress 5.1 WSOD protection.', 'wp-debugging' ),
+			]
+		);
 	}
 
 	/**
@@ -255,6 +299,7 @@ class Settings {
 	private function print_constants() {
 		$constants = [ 'wp_debug_log', 'script_debug', 'savequeries' ];
 		$constants = array_merge( array_keys( self::$options ), $constants );
+		$constants = $this->filter_constants( $constants );
 		echo '<pre>';
 		foreach ( $constants as $constant ) {
 			$value    = 'wp_debug_display' === $constant ? 'false' : 'true';