+++ CHANGES.md
@@ -4,6 +4,7 @@
 * use `wp-cli/wp-config-transformer` to change `wp-config.php`
 * include `norcross/debug-quick-look` as dependency via composer but use my fork
 * update POT via `composer.json` and wp-cli
+* add bypass for WordPress 5.1 WSOD Shutdown Handler
 
 #### 1.2.5 - 1.2.7
 * composer update